refactor: split index.ts to 5 files
