+++ src/controller.ts
@@ -1,0 +1,11 @@
+import { iModel, iView, iController } from './types';
+
+export class Controller implements iController {
+  model: iModel;
+  view: iView;
+
+  constructor(model: iModel, view: iView) {
+    this.model = model;
+    this.view = view;
+  }
+}
