+++ src/model.ts
@@ -1,12 +1,5 @@
-export interface iModel {}
+import { iModel } from './types';
 
-export interface iView {
-  app: Element | null;
-
-  createElement(tag: string, className?: string | null): Element;
-}
-
-export interface iController {
-  model: iModel;
-  view: iView;
+export class Model implements iModel {
+  constructor() {}
 }
